changed response, added exception handlers

+++ compare-cars/src/main/java/com/demo/comparecars/service/impl/CarServiceImpl.java
@@ -1,15 +1,19 @@
 package com.demo.comparecars.service.impl;
 
 import com.demo.comparecars.data.*;
+import com.demo.comparecars.exception.GenericException;
 import com.demo.comparecars.repository.CarRepository;
 import com.demo.comparecars.repository.ScoreConfigRepository;
 import com.demo.comparecars.service.CarService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import javafx.util.Pair;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 
 import java.lang.reflect.Field;
 import java.util.*;
+import java.util.jar.Attributes;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -41,17 +45,22 @@ public class CarServiceImpl implements CarService {
     public void delete(String id) {
         try {
             Car car = carRepository.findById(id).get();
+            if (car == null) {
+                throw new GenericException("Car with given id not found", HttpStatus.NOT_FOUND);
+            }
             carRepository.delete(car);
         } catch (Exception e) {
-//            todo handle
+            throw new GenericException("Car with given id not found", HttpStatus.NOT_FOUND);
         }
 
     }
 
     @Override
     public List<Car> findSimilarCars(String id) {
-//        todo throw exception
         Car car = carRepository.findById(id).get();
+        if (car == null) {
+            throw new GenericException("Car with given id not found", HttpStatus.NOT_FOUND);
+        }
         List<ScoreConfig> all = scoreConfigRepository.findAll();
         HashMap<String, Integer> scoreMap = new HashMap<>();
         for (ScoreConfig each : all) {
@@ -69,7 +78,7 @@ public class CarServiceImpl implements CarService {
                     calculateScoreForBodyType(car, scoreMap, each.getScore());
                     break;
                 default:
-//                    todo throw exception
+                    throw new GenericException("Config Missing", HttpStatus.INTERNAL_SERVER_ERROR);
             }
         }
         TreeSet<CarScore> sortedCars = new TreeSet<>();
@@ -83,6 +92,7 @@ public class CarServiceImpl implements CarService {
         List<Car> carList = sortedCars.stream()
                 .map(CarScore::getCar)
                 .collect(Collectors.toList());
+        carList = carList.subList(0, 10);
         return carList;
     }
 
@@ -115,12 +125,12 @@ public class CarServiceImpl implements CarService {
     }
 
     @Override
-    public List<String> compareCars(CarComparisonRequest request) throws IllegalAccessException {
+    public HashMap<String, List<CarAttribute>> compareCars(CarComparisonRequest request) throws IllegalAccessException {
         HashMap<String, List<CarAttribute>> carAttributes = new HashMap<>();
         Car baseCar = carRepository.findById(request.getBaseCarId()).get();
         log.info("Base car is {}", baseCar);
         List<Car> cars = new ArrayList<>();
-        for (String eachId : request.getOtherIds()){
+        for (String eachId : request.getOtherIds()) {
             Car currentCar = carRepository.findById(eachId).get();
             log.info("current car is {}", currentCar);
             cars.add(currentCar);
@@ -128,10 +138,10 @@ public class CarServiceImpl implements CarService {
 
         List<String> changedProperties = new ArrayList<>();
         for (Car eachCar : cars) {
-            difference(baseCar,eachCar,carAttributes,changedProperties,null);
+            difference(baseCar, eachCar, carAttributes, changedProperties, null, baseCar.getId(), eachCar.getId());
 
         }
-        return changedProperties;
+        return carAttributes;
     }
 
     private final Set<Class> BASE_TYPES = new HashSet(Arrays.asList(
@@ -141,7 +151,7 @@ public class CarServiceImpl implements CarService {
         return BASE_TYPES.contains(clazz);
     }
 
-    private void difference(Object s1, Object s2, HashMap<String, List<CarAttribute>> carAttributes, List<String> changedProperties, String parent) throws IllegalAccessException {
+    private void difference(Object s1, Object s2, HashMap<String, List<CarAttribute>> carAttributes, List<String> changedProperties, String parent, String car1Id, String car2Id) throws IllegalAccessException {
         for (Field field : s1.getClass().getDeclaredFields()) {
             if (parent == null) {
                 parent = s1.getClass().getSimpleName();
@@ -153,17 +163,50 @@ public class CarServiceImpl implements CarService {
                 continue;
             }
             if (value1 == null || value2 == null) {
-                changedProperties.add(parent + "." + field.getName());
+                Pair<String, Object> pair1 = new Pair<>(car1Id, value1);
+                Pair<String, Object> pair2 = new Pair<>(car2Id, value2);
+                addInResponse(carAttributes, parent, field.getName(), pair1, pair2, true);
             } else {
+                Pair<String, Object> pair1 = new Pair<>(car1Id, value1);
+                Pair<String, Object> pair2 = new Pair<>(car2Id, value2);
                 if (isBaseType(value1.getClass())) {
                     if (!Objects.equals(value1, value2)) {
-                        changedProperties.add(parent + "." + field.getName());
+                        addInResponse(carAttributes, parent, field.getName(), pair1, pair2, true);
+                    } else {
+                        addInResponse(carAttributes, parent, field.getName(), pair1, pair2, false);
                     }
                 } else {
-                    difference(value1, value2, carAttributes,changedProperties, parent + "." + field.getName());
+                    difference(value1, value2, carAttributes, changedProperties, parent + "." + field.getName(), car1Id, car2Id);
                 }
-
             }
+        }
+
+    }
+
+    private void addInResponse(HashMap<String, List<CarAttribute>> carAttributes, String parent, String fieldName, Pair baseCar, Pair comparedCar, boolean isDifferent) {
+        if (carAttributes.containsKey(parent)) {
+            List<CarAttribute> attributes = carAttributes.get(parent);
+            if (attributes.stream().filter(o -> o.getName().equals(fieldName)).findFirst().isPresent()) {
+                CarAttribute sameAttribute = attributes.stream().filter(o -> o.getName().equals(fieldName)).findFirst().get();
+                List<Pair> valueList = sameAttribute.getValues();
+                valueList.add(comparedCar);
+
+            } else {
+                List<Pair> valueList = new ArrayList<>();
+                valueList.add(baseCar);
+                valueList.add(comparedCar);
+                CarAttribute carAttribute = CarAttribute.builder().name(fieldName).isDifferent(isDifferent).values(valueList).build();
+                attributes.add(carAttribute);
+                carAttributes.put(parent, attributes);
+            }
+        } else {
+            List<Pair> valueList = new ArrayList<>();
+            valueList.add(baseCar);
+            valueList.add(comparedCar);
+            CarAttribute carAttribute = CarAttribute.builder().name(fieldName).isDifferent(isDifferent).values(valueList).build();
+            List<CarAttribute> attributes = new ArrayList<>();
+            attributes.add(carAttribute);
+            carAttributes.put(parent, attributes);
         }
 
     }
